Revert "Corrected Square Impact point  + texture coords"
This reverts commit b02b87976bb76a9242791cc2089e64f19c782c15.

+++ src/Scene/Scene.cpp
@@ -42,7 +42,7 @@ void Scene::addObject(Object* object) {
 }
 
 Object* Scene::closestObjectIntersected(Ray ray, Point& closestImpact) const {
-    float distance = FLT_MAX;
+    float distance = INT16_MAX;
     Object* Lior = nullptr;
     Point impact;
     bool collisionDetected = false;
